TreeSegmentation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
addict=2.4.0=pypi_0
apex=0.1=pypi_0
autopep8=1.6.0=pyhd3eb1b0_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.5.30=py36h06a4308_0
cudatoolkit=10.2.89=hfd86e86_1
dataclasses=0.8=pyh4f3eec9_6
ffmpeg=4.3=hf484d3e_0
freetype=2.11.0=h70c0345_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
hdf5=1.10.2=hba1933b_1
intel-openmp=2022.0.1=h06a4308_3633
jpeg=9d=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
mmcv-full=1.4.0=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
opencv=3.4.1=py36h6fd60c2_1
opencv-python=4.5.5.62=pypi_0
opencv3=3.1.0=py36_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pypi_0
pillow=8.3.1=py36h2c7a002_0
pip=21.2.2=py36h06a4308_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pyparsing=3.0.7=pypi_0
python=3.6.13=h12debd9_1
pytorch=1.10.2=py3.6_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
setuptools=58.0.4=py36h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=hc218d9a_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.10.2=py36_cu102
torchvision=0.11.3=py36_cu102
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yapf=0.32.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
